src/EasyOCR/requirements.txt
easyocr 
gradio 
lightning 
pandas 
Pillow 
requests 
torch 
